Split stuffs into workspaces to prevent over-compiling.
Plus a shitload of changes.
