+++ .idea/starlight.iml
@@ -2,11 +2,15 @@
 <module type="WEB_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/backend/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/backend/core/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/backend/prisma-cli/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/backend/prisma_lib/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/.tmp" />
       <excludeFolder url="file://$MODULE_DIR$/temp" />
       <excludeFolder url="file://$MODULE_DIR$/tmp" />
       <excludeFolder url="file://$MODULE_DIR$/backend/target" />
+      <excludeFolder url="file://$MODULE_DIR$/backend/prisma-cli" />
+      <excludeFolder url="file://$MODULE_DIR$/backend/prisma_lib" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
