--- backend/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=postgresql://postgres:qu1tg4m3@localhost:5432/testdb
