+++ backend/.gitignore
@@ -2,7 +2,8 @@
 # Edit at https://www.toptal.com/developers/gitignore?templates=rust,rust-analyzer
 
 ### Prisma generated client file ###
-src/prisma.rs
+prisma_lib/src/lib.rs
+.env
 
 ### Rust ###
 # Generated by Cargo
