+++ backend/Cargo.toml
@@ -1,24 +1,13 @@
-[package]
-name = "backend"
-version = "0.1.0"
-edition = "2021"
-license = "AGPL-3"
-authors = [
-    "Tien Dat Pham <dat20036@gmail.com>"
+[workspace]
+members = [
+    "core",
+    "prisma_lib",
+    "prisma-cli"
 ]
-default-run = "backend"
-
-[dependencies]
-rocket = { version = "0.5.1", features = ["json", "secrets"] }
-rocket_okapi = { "version" = "0.9.0", features = ["swagger"] }
-dotenvy = "0.15.7"
-prisma-client-rust = { git = "https://github.com/Brendonovich/prisma-client-rust", tag = "0.6.11", features = ["migrations", "postgresql"] }
-prisma-client-rust-cli = { git = "https://github.com/Brendonovich/prisma-client-rust", tag = "0.6.11", features = ["postgresql"] }
-serde = { version = "1.0", features = ["derive"] }
-ring = "0.17.8"
-hex = "0.4.3"
-chrono = { version = "0.4.39", features = ["clock"] }
-validator = { version = "0.19", features = ["derive"] }
+resolver = "2"
+default-members = [
+    "core"
+]
 
 [profile.release]
 lto = true
