+++ backend/core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "core"
+version = "0.1.0"
+edition = "2021"
+license = "AGPL-3"
+authors = [
+    "Tien Dat Pham <dat20036@gmail.com>"
+]
+
+[dependencies]
+prisma = { path = "../prisma_lib" }
+rocket = { version = "0.5.1", features = ["json", "secrets"] }
+rocket_okapi = { "version" = "0.9.0", features = ["swagger"] }
+serde = { version = "1.0", features = ["derive"] }
+ring = "0.17.8"
+hex = "0.4.3"
+prisma-client-rust = { git = "https://github.com/Brendonovich/prisma-client-rust", tag = "0.6.11", default-features = false, features = ["migrations", "postgresql"] }
+chrono = { version = "0.4.39", features = ["clock"] }
+validator = { version = "0.19", features = ["derive"] }
