+++ backend/core/src/main.rs
@@ -4,11 +4,11 @@ use std::sync::Arc;
 use rocket_okapi::openapi_get_routes;
 use rocket_okapi::swagger_ui::{make_swagger_ui, SwaggerUIConfig};
 use crate::middlewares::cors;
+use prisma;
 
 mod controllers;
 mod tests;
 mod context;
-mod prisma;
 mod utils;
 mod middlewares;
 
@@ -19,9 +19,6 @@ async fn rocket() -> _ {
             .await
             .expect("Failed to create Prisma client"),
     );
-
-    #[cfg(debug_assert)]
-    db._db_push(false).await.unwrap();
 
     rocket::build()
         .attach(cors::Cors)
