+++ backend/core/src/tests/test_index_page.rs
