+++ backend/prisma-cli/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "prisma-cli"
+version = "6.6.6"
+edition = "2021"
+
+[dependencies]
+prisma-client-rust-cli = { git = "https://github.com/Brendonovich/prisma-client-rust", tag = "0.6.11", default-features = false, features = ["postgresql"] }
