+++ backend/prisma_lib/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "prisma"
+version = "6.6.6"
+edition = "2021"
+
+[dependencies]
+prisma-client-rust = { git = "https://github.com/Brendonovich/prisma-client-rust", tag = "0.6.11", default-features = false, features = ["migrations", "postgresql"] }
+serde = "1.0.216"
+
+[lints.rust]
+unused = "allow"
